% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMNlmec.sim.R
\name{SMNlmec.sim}
\alias{SMNlmec.sim}
\title{Generating Censored UNC, DEC, CAR errors with Mixed Effects, for normal, student's-t and slash distribution.}
\usage{
SMNlmec.sim(
  m,
  x,
  z,
  tt,
  nj,
  beta,
  sigma2,
  D,
  phi,
  struc = "UNC",
  typeModel = "Normal",
  p.cens = 0.1,
  n.cens = NULL,
  cens_type = "right",
  nu_set
)
}
\arguments{
\item{m}{Number of individuals.}

\item{x}{Design matrix of the fixed effects of order \code{N x p}, corresponding to vector of fixed effects.}

\item{z}{Design matrix of the random effects of order\code{N x d}, corresponding to vector of random effects.}

\item{tt}{Vector \code{1 x N} with the time the measurements were made, where \code{N} is the total number of measurements for all individuals.}

\item{nj}{Vector \code{1 x m} with the number of measurements of each individual, where \code{m} is the total number of individuals.}

\item{beta}{Vector of values fixed effects.}

\item{sigma2}{Values of the scalar of the variance matrix.}

\item{D}{Variance matrix of the random effects of order \code{d x d}.}

\item{phi}{Vector of parameter in the \code{DEC} and \code{CAR} structure. NULL for \code{UNC}, c(phi_1,phi_2) for \code{DEC} and c(phi_1,1) for \code{CAR}.}

\item{struc}{Structure for the simulated data. Available options are \code{UNC}, \code{DEC} and \code{CAR}.}

\item{typeModel}{Distribution of the simulated data. Available options are \code{Normal}, \code{Student} and \code{Slash}.}

\item{p.cens}{Percentage of censored measurements in the responses. The default value is 0.1.}

\item{n.cens}{Number of censored measurements in the responses. The default value is NULL.}

\item{cens_type}{The direction of cesoring. Available options are \code{left} and \code{right}.}

\item{nu_set}{degrees of freedom for student's-t or slash simulated data. The default value is NULL.}
}
\value{
return list:
\item{cc}{Vector of censoring indicators.}
\item{y_cc}{Vector of responses censoring.}
\item{y_origin}{Vector of original complete responses.}
}
\description{
Generating Censored UNC, DEC, CAR errors with Mixed Effects, for normal, student's-t and slash distribution.
}
