% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMNlmec.S4class.R
\name{SMNlmec.summary-SMNlmecfit-method}
\alias{SMNlmec.summary-SMNlmecfit-method}
\alias{SMNlmec.summary,SMNlmecfit-method}
\title{S4 method for summarizing SMNlmecfit objects}
\description{
S4 method for summarizing SMNlmecfit objects
}
\keyword{internal}
