% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMNlmec.S4class.R
\name{SMNlmec.summary}
\alias{SMNlmec.summary}
\title{SMNlmecfit Summary}
\usage{
SMNlmec.summary(object)
}
\arguments{
\item{object}{An object of class \code{SMNlmecfit}.}
}
\value{
A summary of model estimations, R-hats, standard errors, and criteria.
}
\description{
A generic function to provide a summary for objects of class
\code{SMNlmecfit}.
}
