% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMNlmec.S4class.R
\docType{class}
\name{SMNlmecfit-class}
\alias{SMNlmecfit-class}
\title{Fitted SMNlmec Model Object}
\description{
An S4 class returned by \code{SMNlmec.est()}.
Users usually do not need to create objects of this class directly.
}
\section{Slots}{

\describe{
\item{\code{stan_object}}{stanfit object from rstan.}

\item{\code{model_criteria}}{list, model selection criteria.}

\item{\code{dist_set}}{character, the name of distribution.}

\item{\code{struc_set}}{character, the name of correlation structure.}
}}

