% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMNlmec.S4class.R
\name{SMNlmecfit.creator}
\alias{SMNlmecfit.creator}
\title{Create SMNlmecfit Objects}
\usage{
SMNlmecfit.creator(stan_object, model_criteria, dist_set, struc_set)
}
\arguments{
\item{stan_object}{stanfit object from rstan.}

\item{model_criteria}{list, model selection criteria.}

\item{dist_set}{character, the name of distribution.}

\item{struc_set}{character, the name of correlation structure.}
}
\value{
A SMNlmecfit object containing:
\item{stan_object}{A stanfit object from rstan::stan().}
\item{model_criteria}{A list includes LPML, DIC, EAIC, EBIC, K-L divergence.}
\item{dist_set}{The setting of distribution of the stan model.}
\item{struc_set}{The setting of correlation structure of the stan model.}
}
\description{
A function to create objects of class \code{SMNlmecfit}.
}
