% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTIdata_sub.R
\docType{data}
\name{UTIdata_sub}
\alias{UTIdata_sub}
\title{Sub data set for Unstructured Treatment Interruption Study}
\format{
\describe{
 A data frame with 360 observations on the following 5 variables.
 \item{Patid}{patient ID}
 \item{Days.after.TI}{days after treatment interruption.}
 \item{Fup}{follow-up months}
 \item{RNA}{viral load RNA}
 \item{RNAcens}{censoring indicator for viral load}
}
}
\usage{
data(UTIdata_sub)
}
\description{
Data set from a study of Unstructured Treatment Interruption in HIV-infected adolescents in four institutions in the US. The main outcome is the HIV-1 RNA viral load, which is subject to censoring below the lower limit of detection of the assay (50 copies/mL). The censored observations are indicated by the variable RNAcens. Excluding subjects whose observations are less than 2 and with missing RNA (excluding subject ID C6 T16).
}
\references{
Saitoh, A., Foca, M, et al. (2008), Clinical outcome in perinatally acquired HIV-infected children and adolescents after unstructured treatment interruption, Pediatrics,121, e513-e521.
}
