% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMNlmec.est.R
\name{SMNlmec.est}
\alias{SMNlmec.est}
\title{Bayesian Censored Mixed-Effects Models with Damped Exponential Correlation Structures for Scale Mixture of Normal distributions error}
\usage{
SMNlmec.est(
  N_complete,
  N_cens,
  N_individuals,
  beta_length,
  D_dim,
  x_set,
  z_set,
  tt,
  y_complete,
  y_censor,
  censor_vector,
  nj_vector,
  censor_nj_vector,
  y_ind,
  dist = "Normal",
  struc = "UNC",
  direction = "left",
  thin_num = 5,
  chains_num = 3,
  iter_num = 5000,
  burn_percen = 0.2,
  seed_set = NULL,
  adapt_delta_set = 0.8
)
}
\arguments{
\item{N_complete}{Total number of observations \code{N} in the response vector.}

\item{N_cens}{Total number of censored observations \code{N_cens} in response vector.}

\item{N_individuals}{Total number of subjects \code{n} in the data.}

\item{beta_length}{Value of p in the the \code{1 x p} coefficient vector.}

\item{D_dim}{Dimension of the \code{d x d} variance matrix of the random error.}

\item{x_set}{Design matrix of the fixed effects of order \code{N x p}.}

\item{z_set}{Design matrix of the random effects of order \code{N x d}.}

\item{tt}{Vector \code{1 x N} with the time the measurements were made, where \code{N} is the total number of measurements for all individuals. Default it's considered regular times.}

\item{y_complete}{Vector \code{1 x N} of the complete responses.}

\item{y_censor}{Vector \code{1 x N_cens} of the censored responses in the complete responses.}

\item{censor_vector}{Vector \code{1 x N} of the indicator vector of censored responses.}

\item{nj_vector}{Vector \code{1 x n} of the number of responses for each individual.}

\item{censor_nj_vector}{Vector \code{1 x n} of the number of censored responses for each individual.}

\item{y_ind}{Vector \code{1 x N} to show which subject does a specific response belong to.}

\item{dist}{Distribution of the random effects and random error. Available options are \code{Normal}, \code{Student} and \code{Slash}.}

\item{struc}{Structure of the correlation structure. Available options are \code{UNC}, \code{DEC}, \code{CAR}.}

\item{direction}{Direction of censoring type. Available options are \code{left} and \code{right}.}

\item{thin_num}{A positive integer specifying the period for saving samples. The default is 5. See more details in rstan::stan().}

\item{chains_num}{A positive integer specifying the number of chains generating by rstan::stan(). The default is 3.}

\item{iter_num}{A positive integer specifying the number of iterations for each chain (including warmup). The default is 5000.}

\item{burn_percen}{A percentage of the warm-up iterations in each chain the Stan. The default is 0.2.}

\item{seed_set}{A random seed. The default is NULL.}

\item{adapt_delta_set}{A parameter to control the sampler's behavior. The default is 0.8. See rstan::stan() for more details.}
}
\value{
Return a list of Stan object and a data frame of model selection criteria:
\item{StanObject}{A Stan object from rstan::stan().}
\item{SelectionCriteria}{A data frame of model selection criteria, including LPML, AIC, BIC and DIC.}
}
\description{
This function fits left, right censored mixed-effects linear model, with scale mixture of normal distribution errors, using the Stan. It returns estimates, standard errors and LPML, AIC, BIC and DIC.
}
\examples{
require(lmec) ### lmec is no longer available
require(rstan)
require(StanHeaders)
require(MASS)
require(tmvtnorm)
require(mvtnorm)
require(mnormt)

data("UTIdata")
data1 <- UTIdata
data1 <- subset(data1, !is.na(RNA))
subjects <- unique(data1$Patid)
cluster <- c(match(data1$Patid,subjects))
m <- length(subjects)
N <-length(cluster)
y1 <- c(log10(data1$RNA))
cc = (data1$RNAcens==1)+0
x = cbind((data1$Fup==0)+0, (data1$Fup==1)+0, (data1$Fup==3)+0, (data1$Fup==6)+0, (data1$Fup==9)+0, (data1$Fup==12)+0, (data1$Fup==18)+0, (data1$Fup==24)+0)
tem = data1$Fup
nj<- rep(0,m)
for (j in 1:m) {
  nj[j]=sum(cluster==j)
}

which(nj<3)

############################################
# Excluded subject  8,16,34,55,56          #
############################################

for(i in c(8,16,34,55,56))
{
  y1[cluster==i]= NA
  x[cluster==i]=NA
  tem[cluster==i]=NA
  cc[cluster==i]=NA
  nj[i]=NA
  cluster[cluster==i]=NA
}

paciente=data1$Patid
paciente[paciente=="C6"]=NA
paciente[paciente=="T16"]=NA
paciente=as.vector(na.omit(paciente))


y1=as.vector(na.omit(y1))
x=asRn.Matrix(na.omit(x))
tem=as.vector(na.omit(tem))
cc=as.vector(na.omit(cc))
nj=as.vector(na.omit(nj))
cluster=as.vector(na.omit(cluster))

z <- matrix(rep(1, length(y1)), ncol=1)

###########################################

y_com <- as.numeric(y1)
rho_com <- as.numeric(cc)
ycen<- y_com[rho_com==1]

l_set <- dim(x)[2]
q1_set <- 1

cens_N <- sum(rho_com)
N_com <- length(y_com)
m_com <- m - 5

ind_set <- numeric()
log_nj <- 0
for(i in 1:length(nj)){
  for(j in 1:nj[i]){
    ind_set[log_nj + j] <- i
  }
  log_nj <- log_nj + nj[i]
}

cens_nj <- numeric()
log_nj <- 0
cens_count <- 0

for(i in 1:length(nj)){
  for(j in 1:nj[i]){
    if(rho_com[log_nj + j]==1) {
      cens_count = cens_count + 1
    }
  }
  log_nj = log_nj + nj[i]
  cens_nj[i] = cens_count
  cens_count <- 0
}

SMNlmec_T_DEC <- SMNlmec.est(N_complete = N_com, N_cens = cens_N,
                             N_individuals = m_com, beta_length = l_set,
                             D_dim = q1_set, x_set = x, z_set = z, tt = tem,
                             y_complete = y_com, y_censor = ycen,
                             censor_vector = rho_com, nj_vector = nj,
                             censor_nj_vector = cens_nj,
                             y_ind = ind_set, dist = "Student",
                             struc = "DEC", direction = "left",
                             thin_num = 5, chains_num = 3, iter_num = 5000,
                             burn_percen = 0.2, seed_set = 9955,
                             adapt_delta_set = 0.8)

print(SMNlmec_T_DEC$StanObject,par=c("beta","sigma2","phi1","phi2","D1","nu"), digits = 3)
print(SMNlmec_T_DEC$SelectionCriteria)

}
\references{
TBD
}
