% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMNlmec.summary.R
\docType{methods}
\name{SMNlmec.summary,SMNlmecfit-method}
\alias{SMNlmec.summary,SMNlmecfit-method}
\title{Get a summary of results from SMNlmec.est.}
\usage{
\S4method{SMNlmec.summary}{SMNlmecfit}(object)
}
\arguments{
\item{object}{An object of class \code{SMNlmecfit}.}
}
\value{
The summary of estimations, R hats, standard errors and 95% credible intervals of parameters. Model selection criterion.
}
\description{
Get a summary of results from SMNlmec.est.
}
