% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMNlmec.summary.R
\name{SMNlmec.summary,SMNlmecfit-method}
\alias{SMNlmec.summary,SMNlmecfit-method}
\title{Get a summary of results from SMNlmec.est.}
\usage{
\S4method{SMNlmec.summary}{SMNlmecfit}(object)
}
\arguments{
\item{object}{An object of class \code{SMNlmecfit}.}
}
\description{
Get a summary of results from SMNlmec.est.
}
\examples{
See the example in SMNlmec.est.
}
