% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMNlmec.S4class.R
\docType{class}
\name{SMNlmecfit-class}
\alias{SMNlmecfit-class}
\title{SMNlmecfit Class}
\description{
SMNlmecfit Class
}
\section{Slots}{

\describe{
\item{\code{stan_object}}{stanfit object from rstan.}

\item{\code{model_criteria}}{list, model selection criteria.}

\item{\code{dist_set}}{character, the name of distribution.}

\item{\code{struc_set}}{character, the name of correlation structure.}
}}

