% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMNlmec.S4class.R
\name{SMNlmecfit.creator}
\alias{SMNlmecfit.creator}
\title{Creator Function for SMNlmecfit Class}
\usage{
SMNlmecfit.creator(stan_object, model_criteria, dist_set, struc_set)
}
\arguments{
\item{stan_object}{stanfit object from rstan.}

\item{model_criteria}{list, model selection criteria.}

\item{dist_set}{character, the name of distribution.}

\item{struc_set}{character, the name of correlation structure.}
}
\description{
A function to create objects of class \code{SMNlmecfit}.
}
